Using the steam powered api to return the description
I have implemented a system that sends requests to the steam powered api to return the description for the required game.
The challenges i faced is that i originally was planning to use the steamspypi for the description but within the game dictionary i couldn't find the required information. Furthermore i have plans to clean the HTML that occasionally returns but using Beautiful Soup i couldn't getting  it to work.