Pulling the Game trailer through Steampowered
I have made it so that the game trailer is pulled through a for loop on the steampowered website. Then displayed it on the website. This was done through the use of a for loop and the function .get.

challenges i faced is that the i couldn't originally get it to pull the URL as the ID for it was a really large random number. So i had to implement a for loop to search for the key words webm and MP4 and choose between one. Furthermore an additional challenge is that the app kept displaying every url and all information about the game, but it was a simple fix that took prolonged time as i was returning 1 word 'data' and kept overlooking it. Lastly still haven't solved the issue of it cleaning the html within the description.