Created a system to get every frequently used steam tag
I have created a software that goes to the steam store page popular tags and returns all the most used genres. challenges i have faced are that is i couldn't get it to send a successful request it would always return a failure to connect or too many requests. I also had changed the way i was planning to acquire the genres and that is i was planning to randomly search through each game and pick out genres until there were no more unique genres but it was continuing to fail and just ran into runtime errors.