Fixed video file and changed File and added what platforms game is on
I have fixed the video display sizing so it does not take the whole screen up. I have also removed the HTML within the game description and added a button to show and hide the description. Additionally i am working on searching through a genre and also making a carousel slider for images within the game.
Challenges i am facing is that the how I am going to make them choose genres because because if its an input the user has to be key specific. Additionally I am struggling to add a feature to determine what genre it is being searched through.