Creating the carousel of images and finalized the display
i have added the image slider, moved around buttons and information.
challenges faced where that i originally set out to make a genre search but that plan had quickly failed due to the embedding in html that steam has and my lack of ability to find an endpoint url for each genre with all the games. Additional challenges with the image slider is that it wasn't accepting the trailer and bing search image forcing me to put the aside looking less pleasing and concise, additional issues is that it keeps returning the image to its originall size when transitioning from one image to another and just looks a little poor